Refactored e2e test so it’s easier to read, also fixed issue where CallMethod looked for the last transaction and not the sum of all amounts
